# Performance optimisations: Using more value-types and structs