@Thorium 2023-11-17
The figures are still so bad that people should prefer System.Text.Json and Newtonsoft.Json over this library.