@bartelink 2023-11-17
Yes, far too many moving parts
If going NSJ (which is a bad idea for new systems as it's a dead end), FsCodec.NewtonsoftJson includes required converters for things like options
(FsCodec.SystemTextJson has shims too, but those are a lot less necessary than they are for NSJ)
Also if you want lots of features and dot mind taking a dep on a larger lib, look at FSharp.SystemTextJson