@Thorium 2023-11-17
I asked @eiriktsarpalis the best method for custom parsing json with System.Text.Json, he said:

> just use Utf8JsonReader under wraps. That way make sure it's compliant and also take advantage of all the vectorization code it's using under the hood.

So I tried a bit: https://gist.github.com/Thorium/09a67ce08adee4fcc02ec7f0048e6962
That could work here as well.