@bartelink 2023-11-17
I should mention though - while GC and perf for STJ is in another league, ultimately the perf of JSON serialization is not make or break for 99% of systems. i.e. my motivations for mentioning other libs is not primarily based on perf, more:
- for new systems, you get to make a choice. For older systems, deciding to shift serializers should not be taken lightly
- depending on less semi-maintained code with bespoke behaviors (vs FSharp.Json and FSharp.SystemTextJson)
- STJ and NSJ by having more traffic/being more mainstream is easier to search up answers for on SO and troubleshoot
- NSJ obviously has broad usage and probably most features, but is also extremely a clear dead end - just scan the issue tracker